requirements.txt
Flask==2.3.3
kafka-python==2.0.2
requests==2.31.0
python-dateutil==2.8.2
flask-cors==4.0.0
flask-socketio==5.3.4
python-socketio==5.9.0
redis==5.0.0
elasticsearch==8.11.0
numpy==1.24.3
scikit-learn==1.3.0